move pipeline code into datalad_metalad.pipeline
Moat pipeline related code and data is now located
in the package 'datalad_metalad.pipeline'.
