+++ datalad_metalad/pipeline/consumer/base.py
