+++ datalad_metalad/pipeline/documentedinterface.py
