+++ datalad_metalad/pipeline/pipelinedata.py
@@ -1,6 +1,6 @@
 from copy import deepcopy
-from enum import Enum
 from dataclasses import dataclass, field
+from enum import Enum
 from typing import Any, Dict, Iterable, List, Optional, Tuple
