+++ datalad_metalad/pipeline/pipelines/extract_metadata_autoget_autodrop_pipeline.json
@@ -1,6 +1,6 @@
 {
   "provider": {
-    "module": "datalad_metalad.provider.datasettraverse",
+    "module": "datalad_metalad.pipeline.provider.datasettraverse",
     "class": "DatasetTraverser",
     "name": "traverser",
     "arguments": {}
@@ -8,25 +8,25 @@
   "processors": [
     {
       "name": "autoget",
-      "module": "datalad_metalad.processor.autoget",
+      "module": "datalad_metalad.pipeline.processor.autoget",
       "class": "AutoGet",
       "arguments": {}
     },
     {
       "name": "extractor",
-      "module": "datalad_metalad.processor.extract",
+      "module": "datalad_metalad.pipeline.processor.extract",
       "class": "MetadataExtractor",
       "arguments": {}
     },
     {
       "name": "adder",
-      "module": "datalad_metalad.processor.add",
+      "module": "datalad_metalad.pipeline.processor.add",
       "class": "MetadataAdder",
       "arguments": {}
     },
     {
       "name": "autodrop",
-      "module": "datalad_metalad.processor.autodrop",
+      "module": "datalad_metalad.pipeline.processor.autodrop",
       "class": "AutoDrop",
       "arguments": {}
     }
