+++ datalad_metalad/pipeline/pipelines/extract_metadata_pipeline.json
@@ -1,20 +1,20 @@
 {
   "provider": {
-    "module": "datalad_metalad.provider.datasettraverse",
+    "module": "datalad_metalad.pipeline.provider.datasettraverse",
     "class": "DatasetTraverser",
     "name": "traverser",
     "arguments": {}
   },
   "processors": [
     {
-      "module": "datalad_metalad.processor.extract",
+      "module": "datalad_metalad.pipeline.processor.extract",
       "class": "MetadataExtractor",
       "name": "extractor",
       "arguments": {}
     }
   ],
   "consumer": {
-    "module": "datalad_metalad.consumer.add",
+    "module": "datalad_metalad.pipeline.consumer.add",
     "class": "BatchAdder",
     "name": "adder",
     "arguments": {}
