+++ datalad_metalad/pipeline/processor/base.py
