+++ setup.py
@@ -17,15 +17,16 @@ setup(
     cmdclass=versioneer.get_cmdclass(),
     setup_requires=SETUP_REQUIRES,
     packages=['datalad_metalad',
-              'datalad_metalad.consumer',
+              'datalad_metalad.pipeline',
+              'datalad_metalad.pipeline.consumer',
+              'datalad_metalad.pipeline.processor',
+              'datalad_metalad.pipeline.provider',
               'datalad_metalad.extractors',
               'datalad_metalad.indexers',
               'datalad_metalad.pathutils',
-              'datalad_metalad.processor',
-              'datalad_metalad.provider',
               'datalad_metalad.tests',
               'datalad_metalad.types'],
-    package_data={'datalad_metalad': ['pipelines/*.json']},
+    package_data={'datalad_metalad': ['pipeline/pipelines/*.json']},
     entry_points={
         'datalad.extensions': [
             'metalad=datalad_metalad:command_suite',
